+++ api/separators/bs_roformer/__init__.py
@@ -1,0 +1,3 @@
+from .bs_roformer import BSRoformer
+
+__all__ = ['BSRoformer']