+++ api/tasks.py
@@ -12,12 +12,15 @@ from django.core.files.base import ContentFile
 from django.utils import timezone
 
 from .celery import app
-from .models import (DEMUCS_FAMILY, D3NET, SPLEETER, SPLEETER_PIANO, XUMX,
+from .models import (DEMUCS_FAMILY, D3NET, SPLEETER, SPLEETER_PIANO, XUMX, BS_ROFORMER,
+                     BS_ROFORMER_5S_GUITAR, BS_ROFORMER_5S_PIANO, BS_ROFORMER_6S,
+                     BS_ROFORMER_FAMILY,
                      DynamicMix, SourceFile, StaticMix, TaskStatus,
                      YTAudioDownloadTask)
 from .separators.demucs_separator import DemucsSeparator
 from .separators.spleeter_separator import SpleeterSeparator
-from .util import ALL_PARTS, ALL_PARTS_5, output_format_to_ext, get_valid_filename
+from .separators.bs_roformer_separator import BSRoformerSeparator
+from .util import ALL_PARTS, ALL_PARTS_5_PIANO, ALL_PARTS_5_GUITAR, ALL_PARTS_6, output_format_to_ext, get_valid_filename
 from .youtubedl import download_audio, get_file_ext
 
 """
@@ -37,6 +40,16 @@ def get_separator(separator: str, separator_args: Dict, bitrate: int,
         return SpleeterSeparator(cpu_separation, bitrate, False)
     if separator == SPLEETER_PIANO:
         return SpleeterSeparator(cpu_separation, bitrate, True)
+    if separator in BS_ROFORMER_FAMILY:
+        # Map separator to stem_mode
+        stem_mode_map = {
+            BS_ROFORMER: '4stem',
+            BS_ROFORMER_5S_GUITAR: '5stem_guitar',
+            BS_ROFORMER_5S_PIANO: '5stem_piano',
+            BS_ROFORMER_6S: '6stem',
+        }
+        stem_mode = stem_mode_map.get(separator, '4stem')
+        return BSRoformerSeparator(cpu_separation=cpu_separation, output_format=bitrate, stem_mode=stem_mode)
     if separator in DEMUCS_FAMILY:
         random_shifts = separator_args.get('random_shifts', 0)
         return DemucsSeparator(separator, cpu_separation, bitrate,
@@ -188,7 +201,13 @@ def create_dynamic_mix(dynamic_mix_id):
             dynamic_mix.save()
             return
 
-        all_parts = ALL_PARTS_5 if dynamic_mix.separator == SPLEETER_PIANO else ALL_PARTS
+        all_parts = ALL_PARTS_5_PIANO if dynamic_mix.separator == SPLEETER_PIANO else ALL_PARTS
+        if dynamic_mix.separator == BS_ROFORMER_5S_GUITAR:
+            all_parts = ALL_PARTS_5_GUITAR
+        elif dynamic_mix.separator == BS_ROFORMER_5S_PIANO:
+            all_parts = ALL_PARTS_5_PIANO
+        elif dynamic_mix.separator == BS_ROFORMER_6S:
+            all_parts = ALL_PARTS_6
 
         # Non-local filesystems like S3/Azure Blob do not support source_path()
         is_local = settings.DEFAULT_FILE_STORAGE == 'api.storage.FileSystemStorage'
@@ -357,6 +376,9 @@ def save_to_local_storage(dynamic_mix,
     if 'piano' in parts:
         rel_media_path_piano = os.path.join(
             rel_media_path, f'{file_prefix} (piano) {file_suffix}.{ext}')
+    if 'guitar' in parts:
+        rel_media_path_guitar = os.path.join(
+            rel_media_path, f'{file_prefix} (guitar) {file_suffix}.{ext}')
 
     # File is already on local filesystem
     dynamic_mix.vocals_file.name = rel_media_path_vocals
@@ -365,6 +387,8 @@ def save_to_local_storage(dynamic_mix,
     dynamic_mix.drums_file.name = rel_media_path_drums
     if 'piano' in parts:
         dynamic_mix.piano_file.name = rel_media_path_piano
+    if 'guitar' in parts:
+        dynamic_mix.guitar_file.name = rel_media_path_guitar
 
     dynamic_mix.save()
 
@@ -395,6 +419,8 @@ def save_to_ext_storage(dynamic_mix, rel_path_dir, file_prefix: str,
     dynamic_mix.drums_file = content_files['drums']
     if 'piano' in parts:
         dynamic_mix.piano_file = content_files['piano']
+    if 'guitar' in parts:
+        dynamic_mix.guitar_file = content_files['guitar']
 
     dynamic_mix.save()